# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nakama",
    "description": "REST API aggregating anime, comic, and novel data from 21 public sources behind one consistent JSON interface. GraphQL, AI recommendations, community features, creator dashboard, AI comic generator, CDN edge cache, and auto-generated TypeScript & Python SDKs.",
    "version": "2.8.4"
  },
  "paths": {
...
    "/history": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Record a reading event",
        "operationId": "post_history_history_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List reading history for a user",
        "operationId": "get_history_history_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Numeric user id (service key only)",
              "title": "User Id"
            },
            "description": "Numeric user id (service key only)"
          },
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "anime",
                    "comic",
                    "novel"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by content type (anime/comic/novel)",
              "title": "Content Type"
            },
            "description": "Filter by content type (anime/comic/novel)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response Get History History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryCreate": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Source",
            "examples": [
              "otakudesu"
            ]
          },
          "content_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Content Id",
            "examples": [
              "boruto"
            ]
          },
          "content_type": {
            "type": "string",
            "enum": [
              "anime",
              "comic",
              "novel"
            ],
            "title": "Content Type",
            "examples": [
              "anime"
            ]
          },
          "chapter_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Chapter Id",
            "examples": [
              "episode-1"
            ]
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "Only for service API key callers"
          }
        },
        "type": "object",
        "required": [
          "source",
          "content_id",
          "content_type",
          "chapter_id"
        ],
        "title": "HistoryCreate",
        "description": "Body schema for POST /history."
      },
      "HistoryEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "chapter_id": {
            "type": "string",
            "title": "Chapter Id"
          },
          "read_at": {
            "type": "string",
            "format": "date-time",
            "title": "Read At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "source",
          "content_id",
          "content_type",
          "chapter_id",
          "read_at"
        ],
        "title": "HistoryEntry"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}